extracted methods in Items View Controller
