+++ GuildedRoseLLC/Base.lproj/Main.storyboard
@@ -9,10 +9,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--Greeting View Controller-->
+        <!--Items View Controller-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController storyboardIdentifier="GreetingViewControllerID" id="BYZ-38-t0r" customClass="GreetingViewController" customModule="GuildedRoseLLC" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="GreetingViewControllerID" id="BYZ-38-t0r" customClass="ItemsViewController" customModule="GuildedRoseLLC" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="8bC-Xf-vdC">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
